backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-socketio==5.11.1
aiohttp==3.9.3
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
networkx==3.2.1
redis==5.0.1
elasticsearch==8.11.1
requests==2.31.0
beautifulsoup4==4.12.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru==0.7.2
